+++ firebase-crashlytics/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-
+* [fixed] Execute failure listener outside the main thread  [#6535]
 
 # 19.0.0
 * [fixed] Force validation or rotation of FIDs.
@@ -574,4 +574,3 @@ The following release notes describe changes in the new SDK.
   from your `AndroidManifest.xml` file.
 * [removed] The `fabric.properties` and `crashlytics.properties` files are no
   longer supported. Remove them from your app.
-